evaluator: change test
